allow clicking on entities in graph view in flow run
