+++ apps/hash-frontend/src/pages/[shortname]/shared/flow-visualizer/outputs/entity-result-table.tsx
@@ -15,8 +15,6 @@ import type { EntityRootType, Subgraph } from "@local/hash-subgraph";
 import { Box, TableCell } from "@mui/material";
 import { memo, useMemo, useState } from "react";
 
-import { TypeSlideOverStack } from "../../../../shared/entity-type-page/type-slide-over-stack";
-import { generateEntityRootedSubgraph } from "../../../../shared/subgraphs";
 import type {
   CreateVirtualizedRowContentFn,
   VirtualizedTableColumn,
@@ -27,7 +25,6 @@ import {
   defaultCellSx,
   VirtualizedTable,
 } from "../../../../shared/virtualized-table";
-import { EditEntitySlideOver } from "../../../entities/[entity-uuid].page/edit-entity-slide-over";
 import { EmptyOutputBox } from "./shared/empty-output-box";
 import { outputIcons } from "./shared/icons";
 import { OutputContainer } from "./shared/output-container";
@@ -55,22 +52,12 @@ const columns: VirtualizedTableColumn<FieldId>[] = [
   },
 ];
 
-type ResultSlideOver =
-  | {
-      type: "entity";
-      entity: Entity;
-    }
-  | {
-      type: "entityType";
-      entityTypeId: VersionedUrl;
-    }
-  | null;
-
 type EntityResultRow = {
   entityLabel: string;
   entityTypeId: string;
+  onEntityClick: (entity: Entity) => void;
+  onEntityTypeClick: (entityTypeId: VersionedUrl) => void;
   persistedEntity?: Entity;
-  setSlideOver: (slideOver: ResultSlideOver) => void;
   status: "Proposed" | "New" | "Updated";
 };
 
@@ -90,10 +77,7 @@ const TableRow = memo(({ row }: { row: EntityResultRow }) => {
         <Box
           component="button"
           onClick={() =>
-            row.setSlideOver({
-              type: "entityType",
-              entityTypeId: row.entityTypeId as VersionedUrl,
-            })
+            row.onEntityTypeClick(row.entityTypeId as VersionedUrl)
           }
           sx={{ background: "none", border: "none", p: 0 }}
         >
@@ -116,12 +100,7 @@ const TableRow = memo(({ row }: { row: EntityResultRow }) => {
         {row.persistedEntity ? (
           <Box
             component="button"
-            onClick={() =>
-              row.setSlideOver({
-                type: "entity",
-                entity: row.persistedEntity!,
-              })
-            }
+            onClick={() => row.onEntityClick(row.persistedEntity!)}
             sx={{
               background: "none",
               border: "none",
@@ -147,12 +126,16 @@ const createRowContent: CreateVirtualizedRowContentFn<EntityResultRow> = (
 ) => <TableRow row={row.data} />;
 
 type EntityResultTableProps = {
+  onEntityClick: (entity: Entity) => void;
+  onEntityTypeClick: (entityTypeId: VersionedUrl) => void;
   persistedEntities: PersistedEntity[];
   persistedEntitiesSubgraph?: Subgraph<EntityRootType>;
   proposedEntities: ProposedEntity[];
 };
 
 export const EntityResultTable = ({
+  onEntityClick,
+  onEntityTypeClick,
   persistedEntities,
   persistedEntitiesSubgraph,
   proposedEntities,
@@ -161,8 +144,6 @@ export const EntityResultTable = ({
     field: "entityLabel",
     direction: "asc",
   });
-
-  const [slideOver, setSlideOver] = useState<ResultSlideOver>(null);
 
   const hasData = persistedEntities.length || proposedEntities.length;
 
@@ -211,8 +192,9 @@ export const EntityResultTable = ({
         data: {
           entityLabel,
           entityTypeId,
+          onEntityClick,
+          onEntityTypeClick,
           persistedEntity: "metadata" in entity ? entity : undefined,
-          setSlideOver,
           status:
             "localEntityId" in record
               ? "Proposed"
@@ -229,66 +211,36 @@ export const EntityResultTable = ({
 
       return a.data[field].localeCompare(b.data[field]) * direction;
     });
-  }, [persistedEntities, persistedEntitiesSubgraph, proposedEntities, sort]);
-
-  const selectedEntitySubgraph = useMemo(() => {
-    const defaultEntity = persistedEntities[0]?.entity
-      ? new Entity(persistedEntities[0].entity)
-      : undefined;
-
-    const selectedEntity =
-      slideOver?.type === "entity" ? slideOver.entity : defaultEntity;
-
-    if (!persistedEntitiesSubgraph || !selectedEntity) {
-      return undefined;
-    }
-
-    return generateEntityRootedSubgraph(
-      selectedEntity,
-      persistedEntitiesSubgraph,
-    );
-  }, [slideOver, persistedEntitiesSubgraph]);
+  }, [
+    onEntityClick,
+    onEntityTypeClick,
+    persistedEntities,
+    persistedEntitiesSubgraph,
+    proposedEntities,
+    sort,
+  ]);
 
   return (
-    <>
-      {slideOver?.type === "entityType" && (
-        <TypeSlideOverStack
-          rootTypeId={slideOver.entityTypeId}
-          onClose={() => setSlideOver(null)}
+    <OutputContainer
+      sx={{
+        flex: 1,
+        minWidth: 400,
+      }}
+    >
+      {hasData ? (
+        <VirtualizedTable
+          columns={columns}
+          createRowContent={createRowContent}
+          rows={rows}
+          sort={sort}
+          setSort={setSort}
+        />
+      ) : (
+        <EmptyOutputBox
+          Icon={outputIcons.table}
+          label="Entities proposed and affected by this flow will appear in a table here"
         />
       )}
-      {selectedEntitySubgraph && (
-        <EditEntitySlideOver
-          entitySubgraph={selectedEntitySubgraph}
-          open={slideOver?.type === "entity"}
-          onClose={() => setSlideOver(null)}
-          onSubmit={() => {
-            throw new Error("Editing not permitted in this context");
-          }}
-          readonly
-        />
-      )}
-      <OutputContainer
-        sx={{
-          flex: 1,
-          minWidth: 400,
-        }}
-      >
-        {hasData ? (
-          <VirtualizedTable
-            columns={columns}
-            createRowContent={createRowContent}
-            rows={rows}
-            sort={sort}
-            setSort={setSort}
-          />
-        ) : (
-          <EmptyOutputBox
-            Icon={outputIcons.table}
-            label="Entities proposed and affected by this flow will appear in a table here"
-          />
-        )}
-      </OutputContainer>
-    </>
+    </OutputContainer>
   );
 };
